GH-87: fixed build
