+++ vector/src/main/java/org/apache/arrow/vector/complex/writer/FieldWriter.java
@@ -16,6 +16,7 @@
  */
 package org.apache.arrow.vector.complex.writer;
 
+import org.apache.arrow.vector.complex.writer.BaseWriter.ExtensionWriter;
 import org.apache.arrow.vector.complex.writer.BaseWriter.ListWriter;
 import org.apache.arrow.vector.complex.writer.BaseWriter.MapWriter;
 import org.apache.arrow.vector.complex.writer.BaseWriter.ScalarWriter;
@@ -25,7 +26,8 @@ import org.apache.arrow.vector.complex.writer.BaseWriter.StructWriter;
  * Composite of all writer types. Writers are convenience classes for incrementally adding values to
  * {@linkplain org.apache.arrow.vector.ValueVector}s.
  */
-public interface FieldWriter extends StructWriter, ListWriter, MapWriter, ScalarWriter, ExtensionWriter {
+public interface FieldWriter
+    extends StructWriter, ListWriter, MapWriter, ScalarWriter, ExtensionWriter {
   void allocate();
 
   void clear();
