+++ vector/src/test/java/org/apache/arrow/vector/types/pojo/TestExtensionType.java
@@ -50,6 +50,8 @@ import org.apache.arrow.vector.ipc.ArrowFileReader;
 import org.apache.arrow.vector.ipc.ArrowFileWriter;
 import org.apache.arrow.vector.types.FloatingPointPrecision;
 import org.apache.arrow.vector.types.pojo.ArrowType.ExtensionType;
+import org.apache.arrow.vector.types.pojo.TestUuidVector.UuidType;
+import org.apache.arrow.vector.types.pojo.TestUuidVector.UuidVector;
 import org.apache.arrow.vector.util.VectorBatchAppender;
 import org.apache.arrow.vector.validate.ValidateVectorVisitor;
 import org.junit.jupiter.api.Test;
@@ -292,126 +294,6 @@ public class TestExtensionType {
       VectorBatchAppender.batchAppend(a1, a2, bb);
       assertEquals(6, a1.getValueCount());
       validateVisitor.visit(a1, null);
-    }
-  }
-
-  public static class UuidType extends ExtensionType {
-
-    @Override
-    public ArrowType storageType() {
-      return new ArrowType.FixedSizeBinary(16);
-    }
-
-    @Override
-    public String extensionName() {
-      return "uuid";
-    }
-
-    @Override
-    public boolean extensionEquals(ExtensionType other) {
-      return other instanceof UuidType;
-    }
-
-    @Override
-    public ArrowType deserialize(ArrowType storageType, String serializedData) {
-      if (!storageType.equals(storageType())) {
-        throw new UnsupportedOperationException(
-            "Cannot construct UuidType from underlying type " + storageType);
-      }
-      return new UuidType();
-    }
-
-    @Override
-    public String serialize() {
-      return "";
-    }
-
-    @Override
-    public FieldVector getNewVector(String name, FieldType fieldType, BufferAllocator allocator) {
-      return new UuidVector(name, allocator, new FixedSizeBinaryVector(name, allocator, 16));
-    }
-  }
-
-  public static class UuidVector extends ExtensionTypeVector<FixedSizeBinaryVector>
-      implements ValueIterableVector<UUID> {
-    private final Field field;
-
-    public UuidVector(
-        String name, BufferAllocator allocator, FixedSizeBinaryVector underlyingVector) {
-      super(name, allocator, underlyingVector);
-      this.field = new Field(name, FieldType.nullable(new UuidType()), null);
-    }
-
-    @Override
-    public UUID getObject(int index) {
-      final ByteBuffer bb = ByteBuffer.wrap(getUnderlyingVector().getObject(index));
-      return new UUID(bb.getLong(), bb.getLong());
-    }
-
-    @Override
-    public int hashCode(int index) {
-      return hashCode(index, null);
-    }
-
-    @Override
-    public int hashCode(int index, ArrowBufHasher hasher) {
-      return getUnderlyingVector().hashCode(index, hasher);
-    }
-
-    public void set(int index, UUID uuid) {
-      ByteBuffer bb = ByteBuffer.allocate(16);
-      bb.putLong(uuid.getMostSignificantBits());
-      bb.putLong(uuid.getLeastSignificantBits());
-      getUnderlyingVector().set(index, bb.array());
-    }
-
-    @Override
-    public void copyFromSafe(int fromIndex, int thisIndex, ValueVector from) {
-      getUnderlyingVector()
-          .copyFromSafe(fromIndex, thisIndex, ((UuidVector) from).getUnderlyingVector());
-    }
-
-    @Override
-    public Field getField() {
-      return field;
-    }
-
-    @Override
-    public TransferPair makeTransferPair(ValueVector to) {
-      return new TransferImpl((UuidVector) to);
-    }
-
-    public void setSafe(int index, byte[] value) {
-      getUnderlyingVector().setIndexDefined(index);
-      getUnderlyingVector().setSafe(index, value);
-    }
-
-    public class TransferImpl implements TransferPair {
-      UuidVector to;
-      ValueVector targetUnderlyingVector;
-      TransferPair tp;
-
-      public TransferImpl(UuidVector to) {
-        this.to = to;
-        targetUnderlyingVector = this.to.getUnderlyingVector();
-        tp = getUnderlyingVector().makeTransferPair(targetUnderlyingVector);
-      }
-
-      public UuidVector getTo() {
-        return this.to;
-      }
-
-      public void transfer() {
-        tp.transfer();
-      }
-
-      public void splitAndTransfer(int startIndex, int length) {
-        tp.splitAndTransfer(startIndex, length);
-      }
-
-      public void copyValueSafe(int fromIndex, int toIndex) {
-        tp.copyValueSafe(fromIndex, toIndex);
-      }
     }
   }
